backend/requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
boto==2.49.0
boto3==1.24.63
botocore==1.27.63
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
construct==2.5.3
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-config-url==0.1.1
dj-database-url==1.2.0
Django==4.1.7
django-appconf==1.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.7.0
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==22.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-phonenumber-field==7.0.2
django-rest-framework-nested==0.0.1
django-rest-passwordreset==1.3.0
django-sms==0.6.0
django-storages==1.13.1
django-templated-mail==1.1.1
django-utils-six==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
elastic-transport==8.4.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
elasticsearch6==6.8.2
filelock==3.9.0
future==0.18.2
graphviz==0.20.1
gunicorn==20.1.0
idna==3.4
install==1.3.5
ipaddress==1.0.23
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
pandas==1.5.3
pefile==2022.5.30
phonenumbers==8.13.4
pilkit==2.0
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-ptrace==0.9.8
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.7
pyuca==1.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
twilio==7.16.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
whitenoise==6.4.0
